+++ src/main/java/com/cherrypicker/cherrypick3r/global/aop/baseTimeEntity/domain/BaseTimeEntity.java
@@ -1,4 +1,4 @@
-package com.cherrypicker.cherrypick3r.global.baseTimeEntity.domain;
+package com.cherrypicker.cherrypick3r.global.aop.baseTimeEntity.domain;
 
 import java.time.LocalDateTime;
 import javax.persistence.EntityListeners;
