+++ src/main/java/com/cherrypicker/cherrypick3r/shopPhoto/domain/ShopPhoto.java
@@ -1,6 +1,6 @@
 package com.cherrypicker.cherrypick3r.shopPhoto.domain;
 
-import com.cherrypicker.cherrypick3r.global.baseTimeEntity.domain.BaseTimeEntity;
+import com.cherrypicker.cherrypick3r.global.aop.baseTimeEntity.domain.BaseTimeEntity;
 import com.cherrypicker.cherrypick3r.shop.domain.Shop;
 import com.cherrypicker.cherrypick3r.shopPhoto.dto.ShopPhotoDto;
 import javax.persistence.Column;
