+++ src/main/java/com/cherrypicker/cherrypick3r/userClassify/domain/UserClassify.java
@@ -1,6 +1,6 @@
 package com.cherrypicker.cherrypick3r.userClassify.domain;
 
-import com.cherrypicker.cherrypick3r.global.baseTimeEntity.domain.BaseTimeEntity;
+import com.cherrypicker.cherrypick3r.global.aop.baseTimeEntity.domain.BaseTimeEntity;
 import com.cherrypicker.cherrypick3r.user.domain.User;
 import com.cherrypicker.cherrypick3r.userClassify.dto.UserClassifyDto;
 import java.util.ArrayList;
